requirements.txt
Flask
PyPDF2
pytesseract
pdf2image
Pillow
